Add files via upload
The file FinalApp.py contains the python code to run the streamlit application , where the user gives some random YouTube channel ID as input. The details corresponding to the YouTube channel ID will be displayed when the submit button is clicked.  And the 10 queries are present in the form of a dropdown menu, where user selects one of the query to get the corresponding output. Here the database contains the 10 YouTube Channel ID's and their corresponding data on which the 10 queries will work on to give the output.
The file FinalTABLES_New.ipynb contains the python related to the creation, insertion of the data corresponding to 10 YouTube Channels in the form of Channel_Five Table, Videos present in each channel is stored in the separate table called Video_Data_One table, and the comments of each video present in each youtube channel(maximum 100 comments) is stored in the Comment_Data_Two table. The 10 queries present in the streamlit application works on these tables in order to execute the corresponding usig MySql queries and displays the output.